+++ packages/driver/src/dom/visibility.ts
@@ -147,13 +147,11 @@ const elHasNoEffectiveWidthOrHeight = ($el: JQuery) => {
   // display:none elements, and generally any elements that are not directly rendered,
   // an empty list is returned.
   const el = $el[0]
-  let transform
 
-  if ($el[0].style.transform) {
-    const style = getComputedStyle(el)
+  const style = getComputedStyle(el)
+  let transform = style.getPropertyValue('transform')
 
-    transform = style.getPropertyValue('transform')
-  } else {
+  if (!transform.length) {
     transform = 'none'
   }